Refactor default custom view into CustomViewSample
